Update and rename trnmsr.py to closestWX_faster.py
This is required to generate wxreduce.out for use with a weather bot. Please don't abuse this or my bot to annoy google, i made this because i sometimes really need up to the minute local weather information for various areas around the country, and this seemed the best way to do that. It's all API free, i fetched the list of APRS nodes, and load it into postgres, then run this against that table and use the resulting flatfile for lookups on the fly. The aprs script (elsewhere in this repo) uses beautiful soup 4 to fetch the aprs page and recover the needed information.