Create sah256pad.py
No runtime options yet. No scrypt yet. :-(

But it does work. Append the nonce to the file you give it and the hash will match the hash output.
Adjust "difficulty" to increase the time it takes to find the correct nonce. Not sure how to adjust noncelength for best results yet.