Create aocd4scaffold.py
copilot was lousy with off-by-one misalignments all over the place.
i ended up just using it like stackoverflow, because asking a question and getting a function signature from copilot and then hammering out the implementation is faster than trying to prompt the AI to *just do it* and faster than using python help() or the docs or stack overflow or.

in my opinion.

i'm souring on python, though. i make a lot of typos in code to check code and/or code that looks clever. I don't remember having this many headaches in Java or .net, tbh.